[Mon-07.12-19:01] [DEV] test on mac [NEXT]

+++ philo_one/philo_one.h
@@ -64,7 +64,7 @@ typedef struct	s_data
 	// pthread_mutex_t	stdout_mutex;
 	int				nb_dead_philo;
 	int				param[NB_OF_USEC_ARGS];
-	unsigned int	*last_meal;
+	unsigned long	*last_meal;
 	unsigned int	*nb_meals_eaten;
 }				t_data;
